# api/docs/admin/admin_swagger.yaml
basePath: /api
definitions:
...
  admin.LoginLogIds:
    properties:
      ids:
        items:
          type: integer
        type: array
    required:
    - ids
    type: object
...
  model.AddressBookCollection:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    required:
    - name
    type: object
  model.AddressBookCollectionList:
    properties:
      list:
        items:
          $ref: '#/definitions/model.AddressBookCollection'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  model.LoginLog:
    properties:
      client:
        description: webadmin,webclient,app,
        type: string
      created_at:
        type: string
      device_id:
        type: string
      id:
        type: integer
      ip:
        type: string
      is_deleted:
        type: integer
      platform:
        description: windows,linux,mac,android,ios
        type: string
      type:
        description: account,oauth
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      user_token_id:
        type: integer
      uuid:
        type: string
    type: object
  model.LoginLogList:
    properties:
      list:
        items:
          $ref: '#/definitions/model.LoginLog'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: 接口
  title: 管理系统API
  version: "1.0"
paths:
...
  /admin/config/admin:
    get:
      consumes:
      - application/json
      description: ADMIN服务配置
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: ADMIN服务配置
      tags:
      - ADMIN
  /admin/config/app:
    get:
      consumes:
      - application/json
      description: APP服务配置
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: APP服务配置
      tags:
      - ADMIN
  /admin/config/server:
    get:
      consumes:
      - application/json
      description: 服务配置,给webclient提供api-server
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: LUODA服务配置
      tags:
      - ADMIN
...
  /admin/login_log/batchDelete:
    post:
      consumes:
      - application/json
      description: 登录日志批量删除
      parameters:
      - description: 登录日志
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.LoginLogIds'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 登录日志批量删除
      tags:
      - 登录日志
  /admin/login_log/delete:
    post:
      consumes:
      - application/json
      description: 登录日志删除
      parameters:
      - description: 登录日志信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.LoginLog'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 登录日志删除
      tags:
      - 登录日志
  /admin/login_log/detail/{id}:
    get:
      consumes:
      - application/json
      description: 登录日志详情
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.LoginLog'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 登录日志详情
      tags:
      - 登录日志
  /admin/login_log/list:
    get:
      consumes:
      - application/json
      description: 登录日志列表
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 页大小
        in: query
        name: page_size
        type: integer
      - description: 用户ID
        in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.LoginLogList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 登录日志列表
      tags:
      - 登录日志
...
  /admin/my/address_book_collection/create:
    post:
      consumes:
      - application/json
      description: 创建地址簿名称
      parameters:
      - description: 地址簿名称信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AddressBookCollection'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBookCollection'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 创建地址簿名称
      tags:
      - 我的地址簿名称
  /admin/my/address_book_collection/delete:
    post:
      consumes:
      - application/json
      description: 地址簿名称删除
      parameters:
      - description: 地址簿名称信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AddressBookCollection'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 地址簿名称删除
      tags:
      - 我的地址簿名称
  /admin/my/address_book_collection/list:
    get:
      consumes:
      - application/json
      description: 地址簿名称列表
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 页大小
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBookCollectionList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 地址簿名称列表
      tags:
      - 我的地址簿名称
  /admin/my/address_book_collection/update:
    post:
      consumes:
      - application/json
      description: 地址簿名称编辑
      parameters:
      - description: 地址簿名称信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AddressBookCollection'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBookCollection'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 地址簿名称编辑
      tags:
      - 我的地址簿名称
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  token:
    in: header
    name: api-token
    type: apiKey
swagger: "2.0"